many more pclint warnings fixed

+++ include/uncrustify_types.h
@@ -96,12 +96,12 @@ struct paren_stack_entry_t
 {
    c_token_t    type;         /**< the type that opened the entry */
    size_t       level;        /**< Level of opening type */
-   int          open_line;    /**< line that open symbol is on */
+   size_t       open_line;    /**< line that open symbol is on */
    chunk_t      *pc;          /**< Chunk that opened the level */
-   int          brace_indent; /**< indent for braces - may not relate to indent */
-   int          indent;       /**< indent level (depends on use) */
+   size_t       brace_indent; /**< indent for braces - may not relate to indent */
+   size_t       indent;       /**< indent level (depends on use) */
    size_t       indent_tmp;   /**< temporary indent level (depends on use) */
-   int          indent_tab;   /**< the 'tab' indent (always <= real column) */
+   size_t       indent_tab;   /**< the 'tab' indent (always <= real column) */
    bool         indent_cont;  /**< indent_continue was applied */
    int          ref;
    c_token_t    parent;       /**< if, for, function, etc */
